[#134]Rename FigmaGraph component
